Remove bash dependency from shpec
In 5c49eb4 I added an if [[ ]] check;  this is good practice with bash
but breaks with the dash that Ubuntu (and thus Travis) use by default.
The result is that the remote checks never run, even is ORA_LOGIN is
set: https://travis-ci.org/typekpb/oradown/builds/547074553

Changing to single square brackets [ ] that should work with all POSIX
shells.